requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
idna==3.4
pdfminer.six==20221105
pycparser==2.21
requests==2.31.0
urllib3==2.0.6
